requirements.txt
ics==0.7.2
CairoSVG==2.7.1
pillow
